refactor: global에서 common으로 패키지 변경 (#25)
* test: global에서 common으로 패키지 구조 변경

* test: global로 복귀

* test: ECR_REGISTRY_URI_PREFIX step 안에서 재할당

* test: ECR_REGISTRY_URI_PREFIX을 상위 단계에서 선언

* test: ECR_REGISTRY_URI_PREFIX을 다시 안에서 선언

* test: 디버깅 필드 추가

* test: 디버깅 코드 추가

* test: 수정

* test: 로컬 변수로 수정

* test: 6단계도 로컬 변수로 수정

* test: 기존 서비스 반영하여 업데이트로 수정

* test: 이스케이프 구문 오류 수정

* test: 이스케이프 구문 오류 수정

* test: webhook 테스트

* test: webhook 테스트

* test: webhook 테스트

* test: 브랜치 및 PR 적용

* test: 파일 전략 수정

* chore: 컨테이너 이름이 일치하는 이미지만 찾아 업데이트하도록 변경
