+++ src/main/java/com/couponpop/memberservice/common/config/MonitoringConfig.java
@@ -1,4 +1,4 @@
-package com.couponpop.memberservice.global.config;
+package com.couponpop.memberservice.common.config;
 
 import io.micrometer.core.instrument.MeterRegistry;
 import org.springframework.beans.factory.annotation.Value;
