+++ src/main/java/com/couponpop/memberservice/common/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.couponpop.memberservice.global.config;
+package com.couponpop.memberservice.common.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
