+++ src/main/java/com/couponpop/memberservice/common/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.couponpop.memberservice.global.entity;
+package com.couponpop.memberservice.common.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
