+++ src/main/java/com/couponpop/memberservice/common/exception/CommonErrorCode.java
@@ -1,4 +1,4 @@
-package com.couponpop.memberservice.global.exception;
+package com.couponpop.memberservice.common.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
