+++ src/main/java/com/couponpop/memberservice/common/exception/GlobalExceptionHandler.java
@@ -1,7 +1,7 @@
-package com.couponpop.memberservice.global.exception;
+package com.couponpop.memberservice.common.exception;
 
 
-import com.couponpop.memberservice.global.response.ApiErrorResponse;
+import com.couponpop.memberservice.common.response.ApiErrorResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.support.DefaultMessageSourceResolvable;
