+++ src/main/java/com/couponpop/memberservice/domain/auth/controller/AuthController.java
@@ -1,5 +1,6 @@
 package com.couponpop.memberservice.domain.auth.controller;
 
+import com.couponpop.memberservice.common.response.ApiResponse;
 import com.couponpop.memberservice.domain.auth.dto.request.LoginRequest;
 import com.couponpop.memberservice.domain.auth.dto.request.LogoutRequest;
 import com.couponpop.memberservice.domain.auth.dto.request.SignUpRequest;
@@ -7,7 +8,6 @@ import com.couponpop.memberservice.domain.auth.dto.request.WithdrawRequest;
 import com.couponpop.memberservice.domain.auth.dto.response.LoginResponse;
 import com.couponpop.memberservice.domain.auth.dto.response.SignUpResponse;
 import com.couponpop.memberservice.domain.auth.service.AuthService;
-import com.couponpop.memberservice.global.response.ApiResponse;
 import com.couponpop.security.annotation.CurrentMember;
 import com.couponpop.security.dto.AuthMember;
 import jakarta.validation.Valid;
