+++ src/main/java/com/couponpop/memberservice/domain/auth/service/AuthService.java
@@ -1,6 +1,8 @@
 package com.couponpop.memberservice.domain.auth.service;
 
 import com.couponpop.couponpopcoremodule.dto.fcmtoken.request.FcmTokenExpireRequest;
+import com.couponpop.memberservice.common.client.FcmTokenUserFeignClient;
+import com.couponpop.memberservice.common.exception.GlobalException;
 import com.couponpop.memberservice.domain.auth.dto.request.LoginRequest;
 import com.couponpop.memberservice.domain.auth.dto.request.LogoutRequest;
 import com.couponpop.memberservice.domain.auth.dto.request.SignUpRequest;
@@ -12,8 +14,6 @@ import com.couponpop.memberservice.domain.auth.exception.AuthErrorCode;
 import com.couponpop.memberservice.domain.member.entity.Member;
 import com.couponpop.memberservice.domain.member.exception.MemberErrorCode;
 import com.couponpop.memberservice.domain.member.repository.MemberRepository;
-import com.couponpop.memberservice.global.client.FcmTokenUserFeignClient;
-import com.couponpop.memberservice.global.exception.GlobalException;
 import com.couponpop.security.blacklist.service.TokenBlacklistService;
 import com.couponpop.security.dto.AuthMember;
 import com.couponpop.security.token.JwtProvider;
