+++ src/main/java/com/couponpop/memberservice/domain/member/controller/MemberController.java
@@ -1,9 +1,9 @@
 package com.couponpop.memberservice.domain.member.controller;
 
+import com.couponpop.memberservice.common.response.ApiResponse;
 import com.couponpop.memberservice.domain.member.dto.request.MemberProfileUpdateRequest;
 import com.couponpop.memberservice.domain.member.dto.response.MemberProfileResponse;
 import com.couponpop.memberservice.domain.member.service.MemberService;
-import com.couponpop.memberservice.global.response.ApiResponse;
 import com.couponpop.security.annotation.CurrentMember;
 import com.couponpop.security.dto.AuthMember;
 import jakarta.validation.Valid;
