+++ src/main/java/com/couponpop/memberservice/domain/member/entity/Member.java
@@ -1,8 +1,8 @@
 package com.couponpop.memberservice.domain.member.entity;
 
+import com.couponpop.memberservice.common.entity.BaseEntity;
 import com.couponpop.memberservice.domain.member.dto.request.MemberProfileUpdateRequest;
 import com.couponpop.memberservice.domain.member.enums.MemberType;
-import com.couponpop.memberservice.global.entity.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
