+++ src/main/java/com/couponpop/memberservice/domain/member/exception/MemberErrorCode.java
@@ -1,6 +1,6 @@
 package com.couponpop.memberservice.domain.member.exception;
 
-import com.couponpop.memberservice.global.exception.ErrorCode;
+import com.couponpop.memberservice.common.exception.ErrorCode;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
