+++ src/main/java/com/couponpop/memberservice/domain/member/service/MemberService.java
@@ -1,12 +1,12 @@
 package com.couponpop.memberservice.domain.member.service;
 
+import com.couponpop.memberservice.common.exception.GlobalException;
 import com.couponpop.memberservice.domain.auth.exception.AuthErrorCode;
 import com.couponpop.memberservice.domain.member.dto.request.MemberProfileUpdateRequest;
 import com.couponpop.memberservice.domain.member.dto.response.MemberProfileResponse;
 import com.couponpop.memberservice.domain.member.entity.Member;
 import com.couponpop.memberservice.domain.member.exception.MemberErrorCode;
 import com.couponpop.memberservice.domain.member.repository.MemberRepository;
-import com.couponpop.memberservice.global.exception.GlobalException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
