+++ src/test/java/com/couponpop/memberservice/domain/member/controller/MemberControllerTest.java
@@ -1,12 +1,12 @@
 package com.couponpop.memberservice.domain.member.controller;
 
 
+import com.couponpop.memberservice.common.exception.GlobalException;
 import com.couponpop.memberservice.domain.auth.exception.AuthErrorCode;
 import com.couponpop.memberservice.domain.member.dto.request.MemberProfileUpdateRequest;
 import com.couponpop.memberservice.domain.member.dto.response.MemberProfileResponse;
 import com.couponpop.memberservice.domain.member.enums.MemberType;
 import com.couponpop.memberservice.domain.member.service.MemberService;
-import com.couponpop.memberservice.global.exception.GlobalException;
 import com.couponpop.security.dto.AuthMember;
 import com.couponpop.security.token.JwtAuthFilter;
 import com.couponpop.security.token.JwtAuthenticationToken;
